⚡️: 파일 구조 셋팅
