+++ src/main.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
-import "./index.css";
+import "@/index.css";
 
 const rootElement = document.getElementById("root");
 if (rootElement) {
