+++ src/views/SearchPlace.tsx
@@ -1,0 +1,5 @@
+export default function SearchPlace() {
+  return (
+    <div>SearchPlace</div>
+  )
+}
